update message & restore user api

+++ routes/api.php
@@ -47,7 +47,7 @@ Route::middleware('auth:sanctum')->group(function () {
         Route::get('show/{id}', [UserController::class, 'show'])->middleware('permission:can-user-show');
         Route::post('update', [UserController::class, 'update'])->middleware('permission:can-user-update');
         Route::delete('delete/{id}', [UserController::class, 'destroy'])->middleware('permission:can-user-delete');
-        Route::delete('force-delete/{id}', [UserController::class, 'forceDestroy']);
+        Route::delete('force-delete/{id}', [UserController::class, 'forceDestroy'])->middleware('permission:can-user-delete');
         Route::post('restore', [UserController::class, 'restore']);
         Route::get('details', [UserController::class, 'details']);
     });
